develop:tasks for user

+++ src/main/java/com/marv/taskmanagement/servise/impl/TaskServiceImpl.java
@@ -6,14 +6,19 @@ import com.marv.taskmanagement.model.dto.request.CreateTaskRequest;
 import com.marv.taskmanagement.model.dto.request.UpdateTaskRequest;
 import com.marv.taskmanagement.model.dto.response.TaskResponse;
 import com.marv.taskmanagement.model.entity.TaskEntity;
+import com.marv.taskmanagement.model.entity.UserEntity;
 import com.marv.taskmanagement.model.enums.TaskStatus;
 import com.marv.taskmanagement.repository.TaskRepository;
+import com.marv.taskmanagement.repository.UserRepository;
 import com.marv.taskmanagement.servise.TaskService;
 import lombok.RequiredArgsConstructor;
+import org.apache.catalina.User;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 
@@ -23,11 +28,21 @@ public class TaskServiceImpl implements TaskService {
 
     private final TaskRepository taskRepository;
     private final TaskMapper taskMapper;
+    private final UserRepository userRepository;
+
+    private UserEntity getCurrentUser() {
+        String username = SecurityContextHolder.getContext()
+                .getAuthentication()
+                .getName();
+        return userRepository.findByUsername(username)
+                .orElseThrow(() -> new ResourceNotFoundException("User not found"));
+    }
 
     @Transactional
     @Override
     public TaskResponse create(CreateTaskRequest request) {
         TaskEntity taskEntity = taskMapper.toEntity(request);
+        taskEntity.setUser(getCurrentUser());
         TaskEntity saved = taskRepository.save(taskEntity);
         return taskMapper.toResponse(saved);
     }
@@ -35,27 +50,38 @@ public class TaskServiceImpl implements TaskService {
     @Transactional(readOnly = true)
     @Override
     public Page<TaskResponse> getAll(TaskStatus status, Pageable pageable) {
+        UserEntity currentUser = getCurrentUser();
         if (status != null) {
-            return taskRepository.findByStatus(status, pageable)
+            return taskRepository.findByUserAndStatus(currentUser, status, pageable)
                     .map(taskMapper:: toResponse);
         }
-        return taskRepository.findAll(pageable)
+        return taskRepository.findByUser(currentUser, pageable)
                 .map(taskMapper::toResponse);
     }
 
     @Transactional(readOnly = true)
     @Override
     public TaskResponse getById(Long id) {
+        UserEntity currentUser = getCurrentUser();
         TaskEntity taskEntity = taskRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Task not found with id: " + id));
+
+        if (!taskEntity.getUser().getId().equals(currentUser.getId())) {
+            throw  new ResourceNotFoundException("Task not found with id: " + id);
+        }
         return taskMapper.toResponse(taskEntity);
     }
 
     @Transactional
     @Override
     public TaskResponse update(Long id, UpdateTaskRequest request) {
+        UserEntity currentUser = getCurrentUser();
         TaskEntity existing = taskRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Task not found with id: " + id));
+
+        if (!existing.getUser().getId().equals(currentUser.getId())) {
+            throw new ResourceNotFoundException("Task not found with id: " + id);
+        }
 
         existing.setTitle(request.getTitle());
         existing.setDescription(request.getDescription());
@@ -63,14 +89,19 @@ public class TaskServiceImpl implements TaskService {
         existing.setDeadline(request.getDeadline());
 
         TaskEntity saved = taskRepository.save(existing);
-        return taskMapper.toResponse((saved));
+        return taskMapper.toResponse(saved);
     }
 
     @Transactional
     @Override
     public void delete(Long id) {
-        taskRepository.findById(id)
+        UserEntity currentUser = getCurrentUser();
+        TaskEntity taskEntity = taskRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Task not found with id: " + id));
+
+        if (!taskEntity.getUser().getId().equals(currentUser.getId())) {
+            throw new ResourceNotFoundException("Task not found with id: " + id);
+        }
         taskRepository.deleteById(id);
     }
 }
